# Simplify utf8.c