@am11 2024-05-21
cc @AaronRobinsonMSFT, re: https://github.com/dotnet/runtime/pull/100451#discussion_r1546611006 - this mainly reduces the assembly code with cl.exe -O2 (before 1091 lines, now 369). Could you please run your benchmark to see how it compares against the MultiByteToWideChar/WideCharToMultiChar win32 APIs? Hopefully it will perform better now (at least better than before 😅).